Changed meas mosaic IDS back to 32 bits (as agreed to on telecon)

+++ include/lsst/meas/mosaic/mosaicfit.h
@@ -14,9 +14,9 @@ namespace lsst {
     namespace meas {
 	namespace mosaic {
 
-            typedef boost::int64_t IdType;
+            typedef boost::int32_t IdType;
             typedef boost::int32_t ChipType;
-            typedef boost::int64_t ExpType;
+            typedef boost::int32_t ExpType;
 
             class Source {
             public:
